backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.10
pydantic-settings>=2.7
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
aiosqlite>=0.20
redis>=5.2
python-dotenv>=1.0
httpx>=0.28
tenacity>=9.0
supabase
psycopg[binary]
